requirements.txt
ultralytics>=8.2.0
opencv-python>=4.10.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
onnx>=1.16.0
onnxruntime>=1.18.0
numpy>=1.26.0
pyyaml>=6.0.1
